% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta_arc_extr.r
\name{archive_listing_signature}
\alias{archive_listing_signature}
\title{Construct a stable archive-content signature}
\usage{
archive_listing_signature(x)
}
\arguments{
\item{x}{Archive listing returned by \code{utils::unzip(list = TRUE)}.}
}
\value{
data.table.
}
\description{
Construct a stable archive-content signature
}
\keyword{internal}
