% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta_arc_extr.r
\name{archive_listings_equal}
\alias{archive_listings_equal}
\title{Compare two archive listings}
\usage{
archive_listings_equal(x, y)
}
\arguments{
\item{x, y}{Archive listing tables.}
}
\value{
Logical scalar.
}
\description{
Compare two archive listings
}
\keyword{internal}
