% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta_arc_extr.r
\name{archive_member_exists}
\alias{archive_member_exists}
\title{Test whether an archive member exists in an extraction directory}
\usage{
archive_member_exists(archive_dir, member)
}
\arguments{
\item{archive_dir}{Extraction directory.}

\item{member}{Archive member name.}
}
\value{
Logical scalar.
}
\description{
Test whether an archive member exists in an extraction directory
}
\keyword{internal}
