% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta_arc_extr.r
\name{archive_member_normalise}
\alias{archive_member_normalise}
\title{Normalise an archive member path}
\usage{
archive_member_normalise(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector using forward slashes.
}
\description{
Normalise an archive member path
}
\keyword{internal}
