% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta_arc_extr.r
\name{archive_regex_match}
\alias{archive_regex_match}
\title{Match a regular expression with a structured error}
\usage{
archive_regex_match(x, pattern, argument)
}
\arguments{
\item{x}{Character vector.}

\item{pattern}{Character scalar.}

\item{argument}{Argument name used in an error message.}
}
\value{
Logical vector.
}
\description{
Match a regular expression with a structured error
}
\keyword{internal}
