% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spatial.R
\name{crs_match_terra}
\alias{crs_match_terra}
\title{Match CRS's}
\usage{
crs_match_terra(x, target)
}
\arguments{
\item{x}{Terra spatial object.}

\item{target}{Spatial object with defined target CRS.}
}
\value{
x projected to the CRS of the target. If x or the target have
no or undefined CRS's x is returned as is.
}
\description{
Match the CRS of a target terra object with the CRS of an object with
known CRS
}
