% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip_arc.r
\name{dta_extract}
\alias{dta_extract}
\title{Extract data}
\usage{
dta_extract(
  repo_dir = file.path(path.expand("~"), "Documents"),
  repo = "\\\\.zip$",
  file = NULL,
  sub_file = NULL,
  extract_dir = tempdir()
)
}
\arguments{
\item{repo_dir}{File path indicating where the repository is located. Defaults to the users 'Documents' folder.}

\item{repo}{Keyword or regular expression used to search for the repository.}

\item{file}{Search key (strong) to filter for in the \code{repo_dir}.}

\item{sub_file}{Search key to filter for files in the zip archive to be extracted (i.e., not all contents of a zip archive need to be extracted). If NULL then all contents will be extracted.}

\item{extract_dir}{Directory in which to extract files in the (compressed) data archive. Defalts to \code{temp_dir()}.}
}
\description{
Extract files from a zipped data archive into a temporary (or other) location.
}
\author{
Paul J. Gordijn
}
