% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ekznwr-package.R
\docType{package}
\name{ekznwr-package}
\alias{ekznwr}
\alias{ekznwr-package}
\title{ekznwr: Data management and GIS utilities}
\description{
Tools for scanning, classifying, grouping, locating, and extracting
organisational datasets and geospatial archives.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ekznw/ekznwr}
  \item Report bugs at \url{https://github.com/ekznw/ekznwr/issues}
}

}
\author{
\strong{Maintainer}: Paul Gordijn \email{paul.gordijn@kznwildlife.com}

Authors:
\itemize{
  \item Paul Gordijn \email{paul.gordijn@kznwildlife.com}
}

}
\keyword{internal}
