% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ekznwr-package.R
\name{ekznwr_options}
\alias{ekznwr_options}
\title{Package options}
\description{
\code{ekznwr} uses the following options:
}
\details{
\itemize{
\item \code{ekznwr.extract_dir}: directory used when extracting archives. Defaults
to a session-specific temporary directory.
\item \code{ekznwr.max_map_age}: maximum age in seconds of a cached archive map.
Defaults to 24 hours.
}

Set options with \code{\link[=options]{options()}}, for example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{options(ekznwr.extract_dir = "~/ekznwr_extracts")
}\if{html}{\out{</div>}}
}
