% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_sidecars.R
\name{inventory_assign_groups}
\alias{inventory_assign_groups}
\title{Add inventory sidecar grouping}
\usage{
inventory_assign_groups(dt)
}
\arguments{
\item{dt}{data.table produced by classify_gis_inventory().}
}
\value{
data.table.
}
\description{
Add inventory sidecar grouping
}
