% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_classify.R
\name{inventory_classify_format}
\alias{inventory_classify_format}
\title{Classify inventory file format}
\usage{
inventory_classify_format(ext, path = NULL)
}
\arguments{
\item{ext}{File extension.}

\item{path}{Optional path.}
}
\value{
Character scalar.
}
\description{
Classify inventory file format
}
