% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_classify.R
\name{inventory_classify_role}
\alias{inventory_classify_role}
\title{Classify inventory file role}
\usage{
inventory_classify_role(ext, path = NULL, is_container = FALSE)
}
\arguments{
\item{ext}{File extension.}

\item{path}{Optional file path.}

\item{is_container}{Logical.}
}
\value{
Character scalar.
}
\description{
Classify inventory file role
}
