% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_utils.R
\name{inventory_clean_file_stem}
\alias{inventory_clean_file_stem}
\title{Clean inventory file stems}
\usage{
inventory_clean_file_stem(file_name)
}
\arguments{
\item{file_name}{Character vector of file names.}
}
\value{
Character vector.
}
\description{
Removes known sidecar suffixes to derive a logical dataset stem.
}
