% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_sidecars.R
\name{inventory_detect_sidecar}
\alias{inventory_detect_sidecar}
\title{Detect inventory sidecar files}
\usage{
inventory_detect_sidecar(path)
}
\arguments{
\item{path}{Character vector.}
}
\value{
Logical vector.
}
\description{
Detect inventory sidecar files
}
