% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_utils.R
\name{inventory_file_extension}
\alias{inventory_file_extension}
\title{Derive compound file extension}
\usage{
inventory_file_extension(path)
}
\arguments{
\item{path}{Character vector of file paths.}
}
\value{
Character vector of normalised extensions.
}
\description{
Derive compound file extension
}
