% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_utils.R
\name{inventory_rel_path}
\alias{inventory_rel_path}
\title{Derive a relative path without regex escaping}
\usage{
inventory_rel_path(path, root)
}
\arguments{
\item{path}{Character vector of full paths.}

\item{root}{Character scalar or vector of root paths.}
}
\value{
Character vector of paths relative to root where possible.
}
\description{
Derive a relative path without regex escaping
}
\keyword{internal}
