% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_utils.R
\name{inventory_require_columns}
\alias{inventory_require_columns}
\title{Require columns in an inventory table}
\usage{
inventory_require_columns(dt, required, caller)
}
\arguments{
\item{dt}{A data frame or data.table.}

\item{required}{Character vector of required columns.}

\item{caller}{Calling function name.}
}
\value{
\code{dt}, invisibly.
}
\description{
Require columns in an inventory table
}
\keyword{internal}
