% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_containers.R
\name{is_arcinfo_grid_component}
\alias{is_arcinfo_grid_component}
\title{Detect ArcInfo Grid component}
\usage{
is_arcinfo_grid_component(name, path = name)
}
\arguments{
\item{name}{File name.}

\item{path}{File path.}
}
\value{
Logical vector.
}
\description{
Detect ArcInfo Grid component
}
