% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_containers.R
\name{is_esri_info_component}
\alias{is_esri_info_component}
\title{Detect Esri INFO component}
\usage{
is_esri_info_component(name, path)
}
\arguments{
\item{name}{File name.}

\item{path}{File path.}
}
\value{
Logical vector.
}
\description{
Detect Esri INFO component
}
