% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_containers.R
\name{is_gdb_path}
\alias{is_gdb_path}
\title{Detect Esri File Geodatabase path}
\usage{
is_gdb_path(path)
}
\arguments{
\item{path}{Character vector.}
}
\value{
Logical vector.
}
\description{
Detect Esri File Geodatabase path
}
