% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{read_word_tables}
\alias{read_word_tables}
\title{Read metadata tables from a Word document}
\usage{
read_word_tables(this_file = NULL)
}
\arguments{
\item{this_file}{Path to a Word document.}
}
\value{
A named list of data tables extracted from the document.
}
\description{
Extracts and reshapes tables in a \code{.docx} document.
}
