% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{check_tifs}
\alias{check_tifs}
\title{Check TIFF raster files}
\usage{
check_tifs(tifs = NULL)
}
\arguments{
\item{tifs}{Character vector of TIFF file paths.}
}
\value{
A logical vector indicating whether each file was read
successfully.
}
\description{
Attempts to read and process each supplied TIFF with \code{\link[terra:rast]{terra::rast()}}.
Files that cannot be read are deleted.
}
