% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta_arc_extr.r
\name{dta_arc_extr}
\alias{dta_arc_extr}
\title{Extract archive data}
\usage{
dta_arc_extr(
  parent_dir = file.path(path.expand("~"), "Documents"),
  arc_file = NULL,
  arc_file_filter = NULL
)
}
\arguments{
\item{parent_dir}{File path wherein to search for archived data sets. Defaults to the users 'Documents' folder.}

\item{arc_file}{Search key (string) used to filter by when searching for the archived data in the \code{parent_dir}.}

\item{arc_file_filter}{Search key to filter for files within the (zip) archive to be extracted (i.e., not all contents of the archive need to be extracted). If NULL then all contents will be extracted.}
}
\description{
Extract files from a zipped data archive into a temporary (or other) location.
\preformatted{
**parent_dir** \\
              **|__miscellaneous repo's** \\
                    **|__arc_file** \\
                             **|__arc_file_contents**
}
}
\author{
Paul J. Gordijn
}
