% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta_find.r
\name{dta_find}
\alias{dta_find}
\title{Get data file path}
\usage{
dta_find(
  parent_dir = "~",
  search_key = NULL,
  dir_filter = NULL,
  wanted = NULL,
  unwanted = NULL,
  ...
)
}
\arguments{
\item{parent_dir}{The parent directory within which to search for the data file.}

\item{search_key}{The search key to use to find the exact data file. NBS! This is only applied to the filename, not the directory/path (see \code{dir_filter}). Ideally provide the exact file name of the data file. Else 'regex' wildcards can be used.}

\item{dir_filter}{Parameter used to filter files by directory regex.}

\item{wanted}{Regex search key used to filter \emph{for} wanted files.}

\item{unwanted}{Regex search key used to filter \emph{out} unwanted files.}

\item{...}{Reserved for future use.}
}
\value{
The full path for the specified data set. In addition to regular
files, Esri File Geodatabase (\code{.gdb}) and ArcInfo Binary Grid directories
can be returned. If the data is not found or possible duplicates are
detected, the function will prompt the user accordingly.
}
\description{
Searches within a parent directory for a data file or a
directory-backed Esri dataset given its name and a search key.
}
\author{
Paul J. Gordijn
}
