% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta_fp.r
\name{dta_fp}
\alias{dta_fp}
\title{Generates file path from extracted data}
\usage{
dta_fp(extract_i = NULL, name = NULL, row_n = NULL, last_slash = TRUE)
}
\arguments{
\item{extract_i}{Result from \code{dta_extract()}.}

\item{name}{String used to match the file name of a file from the extracted archive.}

\item{row_n}{Integer specifying the item as presented in the order of the zip_contents table in the extracted archive.}

\item{last_slash}{Removes the last slash, if it exists, on the returned file path. Defaults to \code{TRUE}.}
}
\description{
Uses the result of a successful running of \code{dta_extract()} to generate a file path.
}
\author{
Paul J. Gordijn
}
